Added logging to configuration file reading and CLI #1
